+++ apps/client/src/__tests__/e2e/utils/index.ts
@@ -1,0 +1,7 @@
+/**
+ * E2E Test Utilities
+ * Central export for all test helper functions
+ */
+
+export * from "./editor.utils";
+export * from "./navigation.utils";