+++ apps/client/src/components/Layout/Header/index.tsx
@@ -4,7 +4,7 @@ import Image from "@/components/Image";
 import Notifications from "@/components/Notifications";
 import Search from "@/components/Search";
 import User from "@/components/User";
-import { useScrollbarWidth } from "@/hooks";
+import { useScrollbarWidth } from "@/hooks/useScrollbarWidth";
 import { Link, useLink, usePathname } from "@/router";
 import { Button } from "@atomiton/ui";
 import { useState } from "react";